ICLR 2027 Submissions

+++ experiment_diffusion.py
@@ -63,7 +63,7 @@ def make_basis_params(grid_size, grid_range):
     n_bases = grid_size + 3  # cubic
     h = (grid_range[1] - grid_range[0]) / grid_size
     inv_h = 1.0 / h
-    grid_starts = torch.arange(n_bases).float() * h + grid_range[0] - 3 * h
+    grid_starts = torch.arange(n_bases, dtype=torch.float64) * h + grid_range[0] - 3 * h
     return grid_starts, inv_h, n_bases
 
 
@@ -345,9 +345,9 @@ def main():
 
         # Sample random observation points
         torch.manual_seed(42)
-        obs_x = torch.rand(n_obs, 1) * Lx
-        obs_y = torch.rand(n_obs, 1) * Ly
-        obs_t = torch.rand(n_obs, 1) * T_max
+        obs_x = torch.rand(n_obs, 1, dtype=torch.float64) * Lx
+        obs_y = torch.rand(n_obs, 1, dtype=torch.float64) * Ly
+        obs_t = torch.rand(n_obs, 1, dtype=torch.float64) * T_max
 
         # Get true values at observation points
         obs_u = get_obs(obs_x, obs_y, obs_t)
@@ -409,14 +409,14 @@ def main():
     torch.manual_seed(42)
 
     # Sensor A: x < pi (left half)
-    obs_x_a = torch.rand(n_per_sensor, 1) * math.pi
-    obs_y_a = torch.rand(n_per_sensor, 1) * Ly
-    obs_t_a = torch.rand(n_per_sensor, 1) * T_max
+    obs_x_a = torch.rand(n_per_sensor, 1, dtype=torch.float64) * math.pi
+    obs_y_a = torch.rand(n_per_sensor, 1, dtype=torch.float64) * Ly
+    obs_t_a = torch.rand(n_per_sensor, 1, dtype=torch.float64) * T_max
 
     # Sensor B: x >= pi (right half)
-    obs_x_b = torch.rand(n_per_sensor, 1) * math.pi + math.pi
-    obs_y_b = torch.rand(n_per_sensor, 1) * Ly
-    obs_t_b = torch.rand(n_per_sensor, 1) * T_max
+    obs_x_b = torch.rand(n_per_sensor, 1, dtype=torch.float64) * math.pi + math.pi
+    obs_y_b = torch.rand(n_per_sensor, 1, dtype=torch.float64) * Ly
+    obs_t_b = torch.rand(n_per_sensor, 1, dtype=torch.float64) * T_max
 
     print("  Collecting sensor A observations...")
     u_obs_a = torch.tensor(get_obs(obs_x_a, obs_y_a, obs_t_a),
